stackEden/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
Django==3.0.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
Flask==1.1.2
Flask-Testing==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
nose==1.3.7
pycparser==2.20
pymongo==3.9.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Werkzeug==1.0.1
